+++ src/ESPCompressorFormat.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "esp_compressor/format/esc_format.h"